requirements.txt
PySide6>=6.5.0
python-vlc>=3.0.11
